Cut out SourceContent and mapper from View.
