+++ widget/src/main/java/com/tick/taku/widget/internal/SourceContent.kt
@@ -1,0 +1,5 @@
+package com.tick.taku.widget.internal
+
+internal data class SourceContent(val title: String,
+                                  val description: String,
+                                  val imageUrl: String)
